% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{make_package}
\usage{
make_package(
  source_directory,
  package_directory,
  first_and_last_names,
  github_username,
  email,
  package_description,
  dictionary_file,
  repo_name = "https://cran.seoul.go.kr/"
)
}
\arguments{
\item{source_directory}{character, the full path of the source directory (the directory containing .R files in which work has been done to develop the package)}

\item{package_directory}{character, the full path of the package directory}

\item{first_and_last_names}{character, your first and last names, separated by a space}

\item{github_username}{character, your username on GitHub}

\item{email}{character, your email address}

\item{package_description}{character, description of the package which will be written in \code{DESCRIPTION}}

\item{dictionary_file}{character, the location of the dictionary file (please indicate full file path)}

\item{repo_name}{character, the preferred CRAN repo used to update R packages, Default: 'https://cran.seoul.go.kr/'}
}
\description{
Adds \emph{roxygen2} headers and creates many of the files needed for an R package.
}
\details{
Makes DESCRIPTION and NAMESPACE files.  Handles licensing and documentation.  Creates an roxygen2 header for each function using a custom parameter dictionary.
}
\note{
Adding entries into the custom dictionary file (specified with the \code{dictionary_file} param) will automatically fill in parameters in the roxygen2 header.
}
\examples{
\dontrun{
if(interactive()){
 packagemakeR::make_package(
   source_directory = '~/source-dir/R', 
   package_directory = '~/package-dir',
   first_and_last_names = 'John Doe',
   github_username = 'john-doe',
   email = johndoe@email.com,
   package_description = 'My package description.',
   dictionary_file = '~/git/packagemakeR/inst/extdata/dictionary.R',
   repo_name = 'https://cran.seoul.go.kr/')
 }
}
}
\references{
\href{https://www.r-bloggers.com/2017/05/sinew-a-r-package-to-create-self-populating-roxygen2-skeletons}{R Bloggers}
\cr \href{https://rdrr.io/github/yonicd/sinew/man/makeOxyFile.html}{rdrr.io}
\cr \href{https://github.com/yonicd/sinew}{Sinew GitHub}
}
