% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pretty_headers.R
\name{make_pretty_headers}
\alias{make_pretty_headers}
\title{make_pretty_headers}
\usage{
make_pretty_headers(
  file_name,
  dictionary_file = "~/git/packagemakeR/inst/extdata/dictionary.R"
)
}
\arguments{
\item{file_name}{character, the file name relative to the work directory or, if none, the full path to the file}

\item{dictionary_file}{character, the location of the dictionary file (please indicate full file path)}
}
\description{
Adds \emph{roxygen2} headers to new functions added.
}
\details{
Creates an roxygen2 header for each function using a custom parameter dictionary.
}
\examples{
\dontrun{
if(interactive()){
 packagemakeR::make_pretty_headers(
   file_name = '~/git/demo/my_new_fun.R',
   dictionary_file = '~/git/packagemakeR/inst/extdata/dictionary.R')
 }
}
}
\references{
\href{https://www.r-bloggers.com/2017/05/sinew-a-r-package-to-create-self-populating-roxygen2-skeletons}{R Bloggers}
\cr \href{https://rdrr.io/github/yonicd/sinew/man/makeOxyFile.html}{rdrr.io}
\cr \href{https://github.com/yonicd/sinew}{Sinew GitHub}
}
