% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_package.R
\name{push_package}
\alias{push_package}
\title{push_package}
\usage{
push_package(
  package_directory,
  package_description,
  github_username,
  homepage = "null",
  private = "false",
  fork = "true",
  has_downloads = "true",
  has_issues = "true",
  has_projects = "false",
  has_wiki = "false",
  has_pages = "false"
)
}
\arguments{
\item{package_directory}{character, the full path of the directory containing the package}

\item{package_description}{character, the package description which will be on GitHub}

\item{github_username}{character, your username on GitHub}

\item{homepage}{character, either 'null' or the project URL which will be shown on GitHub as the project homepage, Default: 'null'}

\item{private}{character, a boolean 'true' or 'false' value (but inputted as a character in lowercase letters) which indicates a GitHub repo option, Default: 'false'}

\item{fork}{character, a boolean 'true' or 'false' value (but inputted as a character in lowercase letters) which indicates a GitHub repo option, Default: 'true'}

\item{has_downloads}{character, a boolean 'true' or 'false' value (but inputted as a character in lowercase letters) which indicates a GitHub repo option, Default: 'true'}

\item{has_issues}{character, a boolean 'true' or 'false' value (but inputted as a character in lowercase letters) which indicates a GitHub repo option, Default: 'true'}

\item{has_projects}{character, a boolean 'true' or 'false' value (but inputted as a character in lowercase letters) which indicates a GitHub repo option, Default: 'false'}

\item{has_wiki}{character, a boolean 'true' or 'false' value (but inputted as a character in lowercase letters) which indicates a GitHub repo option, Default: 'false'}

\item{has_pages}{character, a boolean 'true' or 'false' value (but inputted as a character in lowercase letters) which indicates a GitHub repo option, Default: 'false'}
}
\description{
Puts a package on GitHub.
}
\note{
Only tested on Arch Linux.

Your GitHub OAuth token should be in your password store under \code{github/oauth-token}.
}
\examples{
\dontrun{
if(interactive()){
 packagemakeR::push_package(
   package_directory = '~/package-dir', 
   package_description = 'This is the description seen on GitHub.', 
   github_username = 'john-doe',
   homepage = 'null',
   private = 'false',
   fork = 'true',
   has_downloads = 'true',
   has_issues = 'true',
   has_projects = 'false',
   has_wiki = 'false',
   has_pages = 'false')
 }
}
}
